# src/main/resources/static/public/schema/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "0.0.1-SNAPSHOT",
    "title": "Współogarniacz server REST API",
    "contact": {
      "name": "Paulina Szwed",
      "url": "https://github.com/Delebrith",
      "email": "paulina.szwed.1996@gmail.com"
    }
  },
...
  "securityDefinitions": {
    "JWT": {
      "type": "apiKey",
      "in": "header",
      "name": "Authorization"
    }
  },
  "paths": {
...
    "/group/my":{
      "get": {
        "operationId": "getMyGroups",
        "tags": [
          "group"
        ],
        "summary": "Request my groups info",
        "security": [
          {
            "JWT": []
          }
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "If info is found",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Group"
              }
            }
          },
          "400": {
            "description": "If error occurs"
          }
        }
      }
    },
    "/group/search/{size}/{number}":{
      "get": {
        "operationId": "searchGroups",
        "tags": [
          "group"
        ],
        "summary": "Search for groups which names contain given string",
        "security": [
          {
            "JWT": []
          }
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "required": true,
            "type": "string",
            "description": "Part of the searched group name"
          },
          {
            "name": "size",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int32",
            "description": "Size of page of users"
          },
          {
            "name": "number",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int32",
            "description": "Number of page of users"
          }
        ],
        "responses": {
          "200": {
            "description": "If invitation is accepted",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Group"
              }
            }
          },
          "400": {
            "description": "If error occurs"
          }
        }
      }
    },
    "/group/find/{groupId}":{
      "get": {
        "operationId": "getGroup",
        "tags": [
          "group"
        ],
        "summary": "Request group info",
        "security": [
          {
            "JWT": []
          }
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "groupId",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int64",
            "description": "Group id"
          }
        ],
        "responses": {
          "200": {
            "description": "If info is found",
            "schema": {
              "$ref": "#/definitions/Group"
            }
          },
          "404": {
            "description": "If group is not found"
          },
          "400": {
            "description": "If error occurs"
          }
        }
      }
    },
    "/group/create":{
      "post": {
        "operationId": "createGroup",
        "tags": [
          "group"
        ],
        "summary": "Create new group",
        "security": [
          {
            "JWT": []
          }
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "groupDto",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Group"
            },
            "description": "Group details"
          }
        ],
        "responses": {
          "201": {
            "description": "If group is successfully created"
          },
          "400": {
            "description": "If error occurs"
          }
        }
      }
    },
    "/group/{groupId}/invite/{userId}":{
      "post": {
        "operationId": "inviteUserToGroup",
        "tags": [
          "group"
        ],
        "summary": "Invite user to group",
        "security": [
          {
            "JWT": []
          }
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "groupId",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int64",
            "description": "Group id"
          },
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int64",
            "description": "User id"
          }
        ],
        "responses": {
          "202": {
            "description": "If invitation is sent"
          },
          "404": {
            "description": "If group or user is not found"
          },
          "409": {
            "description": "If user already belongs to group"
          },
          "400": {
            "description": "If error occurs"
          }
        }
      }
    },
    "/group/{groupId}/leave":{
      "post": {
        "operationId": "leaveGroup",
        "tags": [
          "group"
        ],
        "summary": "Leave group",
        "security": [
          {
            "JWT": []
          }
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "groupId",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int64",
            "description": "Group id"
          }
        ],
        "responses": {
          "204": {
            "description": "If group is left"
          },
          "404": {
            "description": "If group is not found"
          },
          "409": {
            "description": "If user does not belong to group"
          },
          "400": {
            "description": "If error occurs"
          }
        }
      }
    },
    "/group/{groupId}/join":{
      "post": {
        "operationId": "joinGroup",
        "tags": [
          "group"
        ],
        "summary": "Request permission to join the group",
        "security": [
          {
            "JWT": []
          }
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "groupId",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int64",
            "description": "Group id"
          }
        ],
        "responses": {
          "202": {
            "description": "If request is saved"
          },
          "404": {
            "description": "If group is not found"
          },
          "409": {
            "description": "If user already belongs to group"
          },
          "400": {
            "description": "If error occurs"
          }
        }
      }
    },
    "/group/{groupId}/accept/{userId}":{
      "post": {
        "operationId": "acceptUserIntoGroup",
        "tags": [
          "group"
        ],
        "summary": "Accept user into group",
        "security": [
          {
            "JWT": []
          }
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "groupId",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int64",
            "description": "Group id"
          },
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int64",
            "description": "User id"
          }
        ],
        "responses": {
          "204": {
            "description": "If user is accepted"
          },
          "404": {
            "description": "If group or user is not found"
          },
          "409": {
            "description": "If user is already a member"
          },
          "400": {
            "description": "If error occurs"
          }
        }
      }
    },
    "/group/{groupId}/reject/{userId}":{
      "delete": {
        "operationId": "rejectUserFromGroup",
        "tags": [
          "group"
        ],
        "summary": "Reject user's request to join the group",
        "security": [
          {
            "JWT": []
          }
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "groupId",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int64",
            "description": "Group id"
          },
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int64",
            "description": "User id"
          }
        ],
        "responses": {
          "204": {
            "description": "If user is rejected successfully"
          },
          "404": {
            "description": "If group or user is not found"
          },
          "409": {
            "description": "If user is already a member or was not invited"
          },
          "400": {
            "description": "If error occurs"
          }
        }
      }
    },
    "/group/{groupId}/acceptInvitation":{
      "post": {
        "operationId": "acceptInvitation",
        "tags": [
          "group"
        ],
        "summary": "Accept invitation and join the group",
        "security": [
          {
            "JWT": []
          }
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "groupId",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int64",
            "description": "Group id"
          }
        ],
        "responses": {
          "204": {
            "description": "If invitation is accepted"
          },
          "404": {
            "description": "If group is not found"
          },
          "409": {
            "description": "If user already belongs to group or was not invited"
          },
          "400": {
            "description": "If error occurs"
          }
        }
      }
    },
    "/group/{groupId}/rejectInvitation":{
      "delete": {
        "operationId": "rejectInvitation",
        "tags": [
          "group"
        ],
        "summary": "Reject invitation to the group",
        "security": [
          {
            "JWT": []
          }
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "groupId",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int64",
            "description": "Group id"
          }
        ],
        "responses": {
          "204": {
            "description": "If invitation is rejected successfully"
          },
          "404": {
            "description": "If group is not found"
          },
          "409": {
            "description": "If user already belongs to group or was not invited"
          },
          "400": {
            "description": "If error occurs"
          }
        }
      }
    },
    "/group/{groupId}/task":{
      "post": {
        "operationId": "createTask",
        "tags": [
          "group"
        ],
        "summary": "Create new task and add it to the group",
        "security": [
          {
            "JWT": []
          }
        ],
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "groupId",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int64",
            "description": "Group id"
          },
          {
            "name": "taskInfoDto",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/TaskInfo"
            },
            "description": "Task details"
          }
        ],
        "responses": {
          "201": {
            "description": "If task was created successfully"
          },
          "404": {
            "description": "If group is not found"
          },
          "400": {
            "description": "If error occurs"
          }
        }
      }
    },
    "/group/{groupId}/tasks":{
      "get": {
        "operationId": "getTasks",
        "tags": [
          "group"
        ],
        "summary": "Get group's tasklist",
        "security": [
          {
            "JWT": []
          }
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "groupId",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int64",
            "description": "Group id"
          }
        ],
        "responses": {
          "200": {
            "description": "If tasklist was found successfully",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Task"
              }
            }
          },
          "404": {
            "description": "If group is not found"
          },
          "400": {
            "description": "If error occurs"
          }
        }
      }
    },
    "/group/{groupId}/ranking":{
      "post": {
        "operationId": "createRanking",
        "tags": [
          "group"
        ],
        "summary": "Create new ranking and add it to the group",
        "security": [
          {
            "JWT": []
          }
        ],
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "groupId",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int64",
            "description": "Group id"
          },
          {
            "name": "rankingDto",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Ranking"
            },
            "description": "Task details"
          }
        ],
        "responses": {
          "201": {
            "description": "If ranking was created successfully"
          },
          "404": {
            "description": "If group is not found"
          },
          "400": {
            "description": "If error occurs"
          }
        }
      }
    },
    "/group/{groupId}/rankings":{
      "get": {
        "operationId": "getRankings",
        "tags": [
          "group"
        ],
        "summary": "Get group's rankings list",
        "security": [
          {
            "JWT": []
          }
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "groupId",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int64",
            "description": "Group id"
          }
        ],
        "responses": {
          "200": {
            "description": "If rankings list was found successfully",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Ranking"
              }
            }
          },
          "404": {
            "description": "If group is not found"
          },
          "400": {
            "description": "If error occurs"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "User": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "email": {
          "type": "string",
          "format": "email"
        },
        "password": {
          "type": "string",
          "format": "password"
        },
        "name": {
          "type": "string"
        },
        "authorities": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Authority"
          }
        }
      },
      "required": [
        "email", "name", "password"
      ],
      "example": {
        "email": "user@mail.com",
        "password": "password",
        "name": "User User"
      }
    },
    "Authority": {
      "type": "string",
      "enum": [
        "USER", "ADMIN"
      ]
    },
...
    "Group": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "name": {
          "type": "string"
        },
        "affiliations": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Affiliation"
          }
        }
      },
      "required": [
        "name", "affiliations"
      ],
      "example": {
        "id": 1234,
        "name": "My Group",
        "affiliations": []
      }
    },
    "AffiliationType": {
      "type": "string",
      "enum": [
        "MEMBER", "PENDING_INVITATION", "WAITING_FOR_ACCEPTANCE"
      ]
    },
    "Affiliation": {
      "type": "object",
      "properties": {
        "user": {
          "$ref": "#/definitions/User"
        },
        "state": {
          "$ref": "#/definitions/AffiliationType"
        },
        "lastUpdated": {
          "type": "string",
          "format": "date-time"
        }
      },
      "required": [
        "user", "state"
      ]
    },
    "TaskStatus": {
      "type": "string",
      "enum": [
        "ADDED", "DONE"
      ]
    },
    "Task": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "status": {
          "$ref": "#/definitions/TaskStatus"
        },
        "updateTime": {
          "type": "string",
          "format": "date-time"
        },
        "lastModifiedBy": {
          "$ref": "#/definitions/User"
        }
      },
      "required": [
        "name"
      ]
    },
    "TaskInfo": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        }
      },
      "required": [
        "name"
      ]
    },
    "Ranking": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "name": {
          "type": "string"
        },
        "categories": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Category"
          }
        }
      },
      "required": [
        "name", "participants"
      ]
    },
    "Category": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "name": {
          "type": "string"
        },
        "scores": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Score"
          }
        }
      },
      "required": [
        "name"
      ]
    },
    "Score": {
      "type": "object",
      "properties": {
        "user": {
          "$ref": "#/definitions/User"
        },
        "points": {
          "type": "integer",
          "format": "int32"
        }
      },
      "required": [
        "user", "points"
      ]
    }
  }
}